Standardize on Py::Long type for Python bindings.

+++ src/App/ApplicationPy.cpp
@@ -851,7 +851,7 @@ PyObject *Application::sCheckLinkDepth(PyObject * /*self*/, PyObject *args)
         return nullptr;
 
     PY_TRY {
-        return Py::new_reference_to(Py::Int(GetApplication().checkLinkDepth(depth, MessageOption::Throw)));
+        return Py::new_reference_to(Py::Long(GetApplication().checkLinkDepth(depth, MessageOption::Throw)));
     }PY_CATCH;
 }
 
@@ -926,7 +926,7 @@ PyObject *Application::sSetActiveTransaction(PyObject * /*self*/, PyObject *args
         return nullptr;
 
     PY_TRY {
-        Py::Int ret(GetApplication().setActiveTransaction(name, Base::asBoolean(persist)));
+        Py::Long ret(GetApplication().setActiveTransaction(name, Base::asBoolean(persist)));
         return Py::new_reference_to(ret);
     }PY_CATCH;
 }
@@ -943,7 +943,7 @@ PyObject *Application::sGetActiveTransaction(PyObject * /*self*/, PyObject *args
             Py_Return;
         Py::Tuple ret(2);
         ret.setItem(0,Py::String(name));
-        ret.setItem(1,Py::Int(id));
+        ret.setItem(1,Py::Long(id));
         return Py::new_reference_to(ret);
     }PY_CATCH;
 }
